Added more sprites, sounds, a home menu and improved the game over menu
pretty much everything has its own sprites now, along with sounds when you get an armor pick-up, hit an enemy, and click on one of the buttons when choosing a new power-up upon leveling up. The enemies now also face the player when they are chasing him. A brand new main menu added with a play button and a quit button.  An improved game over menu was also added, with new sprites and a better looking font